Finally fixing these damn test.
Jest runs tests in parallel. Because of this, which by the way, is one of the reasons I picked Jest, we were getting tests adding conflicting data to the database. This resulted in tests sometimes working and sometimes not - it just depended on how tests ran.

Glad that is dealt with.